+++ packages/article-image/__tests__/ios/serializers.js
@@ -1,0 +1,12 @@
+const {
+  addSerializers,
+  compose,
+  flattenStyleTransform,
+  minimalNativeTransform,
+  print
+} = require("@times-components/jest-serializer");
+
+addSerializers(
+  expect,
+  compose(print, flattenStyleTransform, minimalNativeTransform)
+);