+++ packages/article-image/__tests__/web/serializers.js
@@ -1,0 +1,3 @@
+const { addSerializers, rnw } = require("@times-components/jest-serializer");
+
+addSerializers(expect, rnw());