+++ packages/article-list/__tests__/web/serializers.js
@@ -1,0 +1,20 @@
+const {
+  addSerializers,
+  compose,
+  enzymeTreeSerializer,
+  flattenStyleTransform,
+  minimalWebTransform,
+  rnwTransform,
+  rnwPrinter
+} = require("@times-components/jest-serializer");
+
+addSerializers(
+  expect,
+  enzymeTreeSerializer(),
+  compose(
+    rnwPrinter,
+    flattenStyleTransform,
+    minimalWebTransform,
+    rnwTransform()
+  )
+);