+++ packages/article-topics/__tests__/web/serializers.js
@@ -1,0 +1,14 @@
+const {
+  addSerializers,
+  compose,
+  enzymeRenderedSerializer,
+  flattenStyleTransform,
+  minimalWebTransform,
+  print
+} = require("@times-components/jest-serializer");
+
+addSerializers(
+  expect,
+  enzymeRenderedSerializer(),
+  compose(print, flattenStyleTransform, minimalWebTransform)
+);