+++ packages/button/__tests__/ios/serializers.js
@@ -1,0 +1,14 @@
+const {
+  addSerializers,
+  compose,
+  enzymeRenderedSerializer,
+  flattenStyleTransform,
+  minimalNativeTransform,
+  print
+} = require("@times-components/jest-serializer");
+
+addSerializers(
+  expect,
+  enzymeRenderedSerializer(),
+  compose(print, flattenStyleTransform, minimalNativeTransform)
+);