+++ packages/icons/__tests__/web/serializers.js
@@ -1,0 +1,6 @@
+const {
+  addSerializers,
+  minimalWeb
+} = require("@times-components/jest-serializer");
+
+addSerializers(expect, minimalWeb);