+++ packages/jest-serializer/__tests__/ios/flatten-style.test.js
@@ -1,0 +1,7 @@
+import shared from "../shared.flatten";
+
+describe("The flatten style serializer should", () => {
+  describe("for native", () => {
+    shared();
+  });
+});