+++ packages/link/__tests__/web/serializers.js
@@ -1,0 +1,7 @@
+const {
+  addSerializers,
+  enzymeRenderedSerializer,
+  minimalRnw
+} = require("@times-components/jest-serializer");
+
+addSerializers(expect, enzymeRenderedSerializer(), minimalRnw());