+++ packages/markup/__tests__/ios/serializers.js
@@ -1,0 +1,6 @@
+const {
+  addSerializers,
+  flattenStyle
+} = require("@times-components/jest-serializer");
+
+addSerializers(expect, flattenStyle);