+++ packages/markup/__tests__/web/serializers.js
@@ -1,0 +1,12 @@
+const {
+  addSerializers,
+  compose,
+  flattenStyleTransform,
+  rnwTransform,
+  rnwPrinter
+} = require("@times-components/jest-serializer");
+
+addSerializers(
+  expect,
+  compose(rnwPrinter, flattenStyleTransform, rnwTransform())
+);