+++ packages/slice/__tests__/ios/serializers.js
@@ -1,0 +1,7 @@
+const {
+  addSerializers,
+  enzymeRenderedSerializer,
+  flattenStyle
+} = require("@times-components/jest-serializer");
+
+addSerializers(expect, enzymeRenderedSerializer(), flattenStyle);