+++ packages/utils/__tests__/jest.config.js
@@ -1,6 +1,5 @@
 const jestConfigurator = require("@times-components/jest-configurator").default;
 
-module.exports = jestConfigurator(null, __dirname, [
-  "graphql.js",
-  "fetch-gql-schema.js"
-]);
+module.exports = jestConfigurator(null, __dirname, {
+  coverageIgnoreGlobs: ["graphql.js", "fetch-gql-schema.js"]
+});